+++ nequip/data/_dataset/_base_datasets.py
@@ -1,20 +1,13 @@
 import numpy as np
 import logging
-import tempfile
 import inspect
-import functools
 import itertools
 import yaml
 import hashlib
 import math
-from os.path import dirname, basename, abspath
-from typing import Tuple, Dict, Any, List, Callable, Union, Optional, Sequence
-
-import ase
-import ase.io
+from typing import Tuple, Dict, Any, List, Callable, Union, Optional
 
 import torch
-import torch.multiprocessing as mp
 
 from torch_runstats.scatter import scatter_std, scatter_mean
 
@@ -28,13 +21,11 @@ from nequip.data import (
     _NODE_FIELDS,
     _EDGE_FIELDS,
     _GRAPH_FIELDS,
-    _LONG_FIELDS,
 )
 from nequip.utils.batch_ops import bincount
 from nequip.utils.regressor import solver
 from nequip.utils.savenload import atomic_write
-from nequip.utils.multiprocessing import num_tasks
-from .transforms import TypeMapper
+from ..transforms import TypeMapper
 
 
 class AtomicDataset(Dataset):
@@ -669,479 +660,4 @@ class AtomicInMemoryDataset(AtomicDataset):
             # RDF is symmetric
             results[(type2, type1)] = results[(type1, type2)]
 
-        return results
-
-
-class NpzDataset(AtomicInMemoryDataset):
-    """Load data from an npz file.
-
-    To avoid loading unneeded data, keys are ignored by default unless they are in ``key_mapping``, ``include_keys``,
-    or ``npz_fixed_fields_keys``.
-
-    Args:
-        key_mapping (Dict[str, str]): mapping of npz keys to ``AtomicData`` keys. Optional
-        include_keys (list): the attributes to be processed and stored. Optional
-        npz_fixed_field_keys: the attributes that only have one instance but apply to all frames. Optional
-            Note that the mapped keys (as determined by the _values_ in ``key_mapping``) should be used in
-            ``npz_fixed_field_keys``, not the original npz keys from before mapping. If an npz key is not
-            present in ``key_mapping``, it is mapped to itself, and this point is not relevant.
-
-    Example: Given a npz file with 10 configurations, each with 14 atoms.
-
-        position: (10, 14, 3)
-        force: (10, 14, 3)
-        energy: (10,)
-        Z: (14)
-        user_label1: (10)        # per config
-        user_label2: (10, 14, 3) # per atom
-
-    The input yaml should be
-
-    ```yaml
-    dataset: npz
-    dataset_file_name: example.npz
-    include_keys:
-      - user_label1
-      - user_label2
-    npz_fixed_field_keys:
-      - cell
-      - atomic_numbers
-    key_mapping:
-      position: pos
-      force: forces
-      energy: total_energy
-      Z: atomic_numbers
-    graph_fields:
-      - user_label1
-    node_fields:
-      - user_label2
-    ```
-
-    """
-
-    def __init__(
-        self,
-        root: str,
-        key_mapping: Dict[str, str] = {
-            "positions": AtomicDataDict.POSITIONS_KEY,
-            "energy": AtomicDataDict.TOTAL_ENERGY_KEY,
-            "force": AtomicDataDict.FORCE_KEY,
-            "forces": AtomicDataDict.FORCE_KEY,
-            "Z": AtomicDataDict.ATOMIC_NUMBERS_KEY,
-            "atomic_number": AtomicDataDict.ATOMIC_NUMBERS_KEY,
-        },
-        include_keys: List[str] = [],
-        npz_fixed_field_keys: List[str] = [],
-        file_name: Optional[str] = None,
-        url: Optional[str] = None,
-        AtomicData_options: Dict[str, Any] = {},
-        include_frames: Optional[List[int]] = None,
-        type_mapper: TypeMapper = None,
-    ):
-        self.key_mapping = key_mapping
-        self.npz_fixed_field_keys = npz_fixed_field_keys
-        self.include_keys = include_keys
-
-        super().__init__(
-            file_name=file_name,
-            url=url,
-            root=root,
-            AtomicData_options=AtomicData_options,
-            include_frames=include_frames,
-            type_mapper=type_mapper,
-        )
-
-    @property
-    def raw_file_names(self):
-        return [basename(self.file_name)]
-
-    @property
-    def raw_dir(self):
-        return dirname(abspath(self.file_name))
-
-    def get_data(self):
-
-        data = np.load(self.raw_dir + "/" + self.raw_file_names[0], allow_pickle=True)
-
-        # only the keys explicitly mentioned in the yaml file will be parsed
-        keys = set(list(self.key_mapping.keys()))
-        keys.update(self.npz_fixed_field_keys)
-        keys.update(self.include_keys)
-        keys = keys.intersection(set(list(data.keys())))
-
-        mapped = {self.key_mapping.get(k, k): data[k] for k in keys}
-
-        for intkey in _LONG_FIELDS:
-            if intkey in mapped:
-                mapped[intkey] = mapped[intkey].astype(np.int64)
-
-        fields = {k: v for k, v in mapped.items() if k not in self.npz_fixed_field_keys}
-        num_examples, num_atoms, n_dim = fields[AtomicDataDict.POSITIONS_KEY].shape
-        assert n_dim == 3
-
-        # now we replicate and add the fixed fields:
-        for fixed_field in self.npz_fixed_field_keys:
-            orig = mapped[fixed_field]
-            if fixed_field in _NODE_FIELDS:
-                assert orig.ndim >= 1  # [n_atom, feature_dims]
-                assert orig.shape[0] == num_atoms
-                replicated = np.expand_dims(orig, 0)
-                replicated = np.tile(
-                    replicated,
-                    (num_examples,) + (1,) * len(replicated.shape[1:]),
-                )  # [n_example, n_atom, feature_dims]
-            elif fixed_field in _GRAPH_FIELDS:
-                # orig is [feature_dims]
-                replicated = np.expand_dims(orig, 0)
-                replicated = np.tile(
-                    replicated,
-                    (num_examples,) + (1,) * len(replicated.shape[1:]),
-                )  # [n_example, feature_dims]
-            else:
-                raise KeyError(
-                    f"npz_fixed_field_keys contains `{fixed_field}`, but it isn't registered as a node or graph field"
-                )
-            fields[fixed_field] = replicated
-        return fields
-
-
-def _ase_dataset_reader(
-    rank: int,
-    world_size: int,
-    tmpdir: str,
-    ase_kwargs: dict,
-    atomicdata_kwargs: dict,
-    include_frames,
-    global_options: dict,
-) -> Union[str, List[AtomicData]]:
-    """Parallel reader for all frames in file."""
-    if world_size > 1:
-        from nequip.utils._global_options import _set_global_options
-
-        # ^ avoid import loop
-        # we only `multiprocessing` if world_size > 1
-        _set_global_options(global_options)
-    # interleave--- in theory it is better for performance for the ranks
-    # to read consecutive blocks, but the way ASE is written the whole
-    # file gets streamed through all ranks anyway, so just trust the OS
-    # to cache things sanely, which it will.
-    # ASE handles correctly the case where there are no frames in index
-    # and just gives an empty list, so that will succeed:
-    index = slice(rank, None, world_size)
-    if include_frames is None:
-        # count includes 0, 1, ..., inf
-        include_frames = itertools.count()
-
-    datas = []
-    # stream them from ase too using iread
-    for i, atoms in enumerate(ase.io.iread(**ase_kwargs, index=index, parallel=False)):
-        global_index = rank + (world_size * i)
-        datas.append(
-            (
-                global_index,
-                AtomicData.from_ase(atoms=atoms, **atomicdata_kwargs)
-                if global_index in include_frames
-                # in-memory dataset will ignore this later, but needed for indexing to work out
-                else None,
-            )
-        )
-    # Save to a tempfile---
-    # there can be a _lot_ of tensors here, and rather than dealing with
-    # the complications of running out of file descriptors and setting
-    # sharing methods, since this is a one time thing, just make it simple
-    # and avoid shared memory entirely.
-    if world_size > 1:
-        path = f"{tmpdir}/rank{rank}.pth"
-        torch.save(datas, path)
-        return path
-    else:
-        return datas
-
-
-class ASEDataset(AtomicInMemoryDataset):
-    """
-
-    Args:
-        ase_args (dict): arguments for ase.io.read
-        include_keys (list): in addition to forces and energy, the keys that needs to
-             be parsed into dataset
-             The data stored in ase.atoms.Atoms.array has the lowest priority,
-             and it will be overrided by data in ase.atoms.Atoms.info
-             and ase.atoms.Atoms.calc.results. Optional
-        key_mapping (dict): rename some of the keys to the value str. Optional
-
-    Example: Given an atomic data stored in "H2.extxyz" that looks like below:
-
-    ```H2.extxyz
-    2
-    Properties=species:S:1:pos:R:3 energy=-10 user_label=2.0 pbc="F F F"
-     H       0.00000000       0.00000000       0.00000000
-     H       0.00000000       0.00000000       1.02000000
-    ```
-
-    The yaml input should be
-
-    ```
-    dataset: ase
-    dataset_file_name: H2.extxyz
-    ase_args:
-      format: extxyz
-    include_keys:
-      - user_label
-    key_mapping:
-      user_label: label0
-    chemical_symbols:
-      - H
-    ```
-
-    for VASP parser, the yaml input should be
-    ```
-    dataset: ase
-    dataset_file_name: OUTCAR
-    ase_args:
-      format: vasp-out
-    key_mapping:
-      free_energy: total_energy
-    chemical_symbols:
-      - H
-    ```
-
-    """
-
-    def __init__(
-        self,
-        root: str,
-        ase_args: dict = {},
-        file_name: Optional[str] = None,
-        url: Optional[str] = None,
-        AtomicData_options: Dict[str, Any] = {},
-        include_frames: Optional[List[int]] = None,
-        type_mapper: TypeMapper = None,
-        key_mapping: Optional[dict] = None,
-        include_keys: Optional[List[str]] = None,
-    ):
-        self.ase_args = {}
-        self.ase_args.update(getattr(type(self), "ASE_ARGS", dict()))
-        self.ase_args.update(ase_args)
-        assert "index" not in self.ase_args
-        assert "filename" not in self.ase_args
-
-        self.include_keys = include_keys
-        self.key_mapping = key_mapping
-
-        super().__init__(
-            file_name=file_name,
-            url=url,
-            root=root,
-            AtomicData_options=AtomicData_options,
-            include_frames=include_frames,
-            type_mapper=type_mapper,
-        )
-
-    @classmethod
-    def from_atoms_list(cls, atoms: Sequence[ase.Atoms], **kwargs):
-        """Make an ``ASEDataset`` from a list of ``ase.Atoms`` objects.
-
-        If `root` is not provided, a temporary directory will be used.
-
-        Please note that this is a convinience method that does NOT avoid a round-trip to disk; the provided ``atoms`` will be written out to a file.
-
-        Ignores ``kwargs["file_name"]`` if it is provided.
-
-        Args:
-            atoms
-            **kwargs: passed through to the constructor
-        Returns:
-            The constructed ``ASEDataset``.
-        """
-        if "root" not in kwargs:
-            tmpdir = tempfile.TemporaryDirectory()
-            kwargs["root"] = tmpdir.name
-        else:
-            tmpdir = None
-        kwargs["file_name"] = tmpdir.name + "/atoms.xyz"
-        atoms = list(atoms)
-        # Write them out
-        ase.io.write(kwargs["file_name"], atoms, format="extxyz")
-        # Read them in
-        obj = cls(**kwargs)
-        if tmpdir is not None:
-            # Make it keep a reference to the tmpdir to keep it alive
-            # When the dataset is garbage collected, the tmpdir will
-            # be too, and will (hopefully) get deleted eventually.
-            # Or at least by end of program...
-            obj._tmpdir_ref = tmpdir
-        return obj
-
-    @property
-    def raw_file_names(self):
-        return [basename(self.file_name)]
-
-    @property
-    def raw_dir(self):
-        return dirname(abspath(self.file_name))
-
-    def get_data(self):
-        ase_args = {"filename": self.raw_dir + "/" + self.raw_file_names[0]}
-        ase_args.update(self.ase_args)
-
-        # skip the None arguments
-        kwargs = dict(
-            include_keys=self.include_keys,
-            key_mapping=self.key_mapping,
-        )
-        kwargs = {k: v for k, v in kwargs.items() if v is not None}
-        kwargs.update(self.AtomicData_options)
-        n_proc = num_tasks()
-        with tempfile.TemporaryDirectory() as tmpdir:
-            from nequip.utils._global_options import _get_latest_global_options
-
-            # ^ avoid import loop
-            reader = functools.partial(
-                _ase_dataset_reader,
-                world_size=n_proc,
-                tmpdir=tmpdir,
-                ase_kwargs=ase_args,
-                atomicdata_kwargs=kwargs,
-                include_frames=self.include_frames,
-                # get the global options of the parent to initialize the worker correctly
-                global_options=_get_latest_global_options(),
-            )
-            if n_proc > 1:
-                # things hang for some obscure OpenMP reason on some systems when using `fork` method
-                ctx = mp.get_context("forkserver")
-                with ctx.Pool(processes=n_proc) as p:
-                    # map it over the `rank` argument
-                    datas = p.map(reader, list(range(n_proc)))
-                # clean up the pool before loading the data
-                datas = [torch.load(d) for d in datas]
-                datas = sum(datas, [])
-                # un-interleave the datas
-                datas = sorted(datas, key=lambda e: e[0])
-            else:
-                datas = reader(rank=0)
-                # datas here is already in order, stride 1 start 0
-                # no need to un-interleave
-        # return list of AtomicData:
-        return [e[1] for e in datas]
-
-
-class HDF5Dataset(AtomicDataset):
-    def __init__(
-        self,
-        root: str,
-        file_name: Optional[str] = None,
-        extra_fixed_fields: Dict[str, Any] = {},
-        type_mapper: Optional[TypeMapper] = None,
-    ):
-        super().__init__(root=root, type_mapper=type_mapper)
-        self.file_name = file_name
-        self.r_max = extra_fixed_fields["r_max"]
-        self.index = None
-        self.num_molecules = 0
-        import h5py
-
-        files = [h5py.File(f, "r") for f in self.file_name.split(";")]
-        for file in files:
-            for group_name in file:
-                self.num_molecules += len(file[group_name]["energy"])
-            file.close()
-
-    def setup_index(self):
-        import h5py
-
-        files = [h5py.File(f, "r") for f in self.file_name.split(";")]
-        self.has_forces = False
-        self.index = []
-        for file in files:
-            for group_name in file:
-                group = file[group_name]
-                types = np.array(group["types"])
-                pos = np.array(group["pos"])
-                energy = np.array(group["energy"])
-                if "forces" in group:
-                    self.has_forces = True
-                    forces = -np.array(group["forces"])
-                    for i in range(len(energy)):
-                        self.index.append((types, pos, energy, forces, i))
-                else:
-                    for i in range(len(energy)):
-                        self.index.append((types, pos, energy, i))
-
-    def len(self) -> int:
-        return self.num_molecules
-
-    def get(self, idx: int) -> AtomicData:
-        if self.index is None:
-            self.setup_index()
-        data = self.index[idx]
-        i = data[-1]
-        args = {
-            "pos": data[1][i],
-            "r_max": self.r_max,
-            AtomicDataDict.ATOM_TYPE_KEY: data[0][i],
-            AtomicDataDict.TOTAL_ENERGY_KEY: data[2][i],
-        }
-        if self.has_forces:
-            args[AtomicDataDict.FORCE_KEY] = data[3][i]
-        return AtomicData.from_points(**args)
-
-    def statistics(
-        self,
-        fields: List[Union[str, Callable]],
-        modes: List[str],
-        stride: int = 1,
-        unbiased: bool = True,
-        kwargs: Optional[Dict[str, dict]] = {},
-    ) -> List[tuple]:
-        assert len(modes) == len(fields)
-        if len(fields) == 0:
-            return []
-        if self.index is None:
-            self.setup_index()
-        results = []
-        for field, mode in zip(fields, modes):
-            count = 0
-            if mode == "rms":
-                total = 0.0
-            elif mode in ("mean_std", "per_atom_mean_std"):
-                total = [0.0, 0.0]
-            else:
-                raise NotImplementedError(f"Analysis mode '{mode}' is not implemented")
-            for index in range(0, self.len(), stride):
-                data = self.index[index]
-                i = data[-1]
-                if field == AtomicDataDict.FORCE_KEY:
-                    values = data[3][i]
-                elif field == AtomicDataDict.TOTAL_ENERGY_KEY:
-                    values = data[2][i]
-                elif callable(field):
-                    values, _ = field(self.get(index))
-                    values = np.asarray(values)
-                else:
-                    raise RuntimeError(
-                        f"The field key `{field}` is not present in this dataset"
-                    )
-                if mode == "rms":
-                    total += np.sum(values * values)
-                    count += len(values.flatten())
-                else:
-                    length = len(values.flatten())
-                    if mode == "per_atom_mean_std":
-                        values /= len(data[0][i])
-                    sample_mean = np.mean(values)
-                    new_mean = (total[0] * count + sample_mean * length) / (
-                        count + length
-                    )
-                    total[1] += (
-                        length * (sample_mean - total[0]) * (sample_mean - new_mean)
-                    )
-                    total[0] = new_mean
-                    count += length
-            if mode == "rms":
-                results.append(torch.tensor((np.sqrt(total / count),)))
-            else:
-                results.append(
-                    (torch.tensor(total[0]), torch.tensor(np.sqrt(total[1] / count)))
-                )
         return results
